Extract debug Monaco samples to lib
